LiveKit/LiveKit/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
av==14.4.0
azure-core==1.34.0
azure-identity==1.23.0
azure-storage-blob==12.25.1
backoff==2.2.1
bcrypt==4.3.0
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.12
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.3
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
docopt==0.6.2
docstring_parser==0.16
dotenv==0.9.9
durationpy==0.10
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.115.9
filelock==3.18.0
firebase-admin==6.9.0
flatbuffers==25.2.10
frozenlist==1.6.2
fsspec==2025.5.1
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gotrue==2.12.0
groq==0.26.0
grpcio==1.72.1
grpcio-status==1.62.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.3
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.64
langchain-groq==0.3.2
langchain-mistralai==0.2.10
langchain-text-splitters==0.3.8
langsmith==0.3.45
livekit==1.0.8
livekit-agents==1.0.23
livekit-api==1.0.2
livekit-plugins-deepgram==1.0.23
livekit-plugins-elevenlabs==1.0.23
livekit-plugins-groq==1.0.23
livekit-plugins-openai==1.0.23
livekit-plugins-silero==1.0.23
livekit-protocol==1.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
msgpack==1.1.0
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
num2words==0.5.14
numpy==2.2.6
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.84.0
opentelemetry-api==1.34.0
opentelemetry-exporter-otlp-proto-common==1.34.0
opentelemetry-exporter-otlp-proto-grpc==1.34.0
opentelemetry-instrumentation==0.55b0
opentelemetry-instrumentation-asgi==0.55b0
opentelemetry-instrumentation-fastapi==0.55b0
opentelemetry-proto==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
opentelemetry-util-http==0.55b0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pillow==11.2.1
pluggy==1.6.0
postgrest==1.0.2
posthog==4.4.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.26.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
SQLAlchemy==2.0.41
starlette==0.45.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.2
supafunc==0.9.4
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
typer==0.16.0
types-protobuf==4.25.0.20240417
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
